+++ src/app/gui/show-code-dialog/tabs/angular-tab.component.ts
@@ -1,6 +1,7 @@
 import {Component, Input} from '@angular/core'
-import {Params} from "../../../model/params";
+import {ComponentParams} from "../../../model/component-params";
 import {copyToClipboard} from "../../../shared/utils";
+import packageInfo  from '../../../../../package.json'
 
 @Component({
   selector: 'app-angular-tab',
@@ -8,13 +9,13 @@ import {copyToClipboard} from "../../../shared/utils";
 })
 export class AngularTabComponent {
 
-  @Input() data: Params
+  @Input() data: ComponentParams
   public link1: string = "npm install animated-banner-component --save"
   public link2: string = `import {defineCustomElements} from "animated-banner/loader"
 ...
 void defineCustomElements(window)`
   public get link3(): string {
-    return `<script type='module' src='https://unpkg.com/browse/animated-banner@0.0.10/dist/animations/${this.data.animation}.js'></script>`
+    return `<script type='module' src='https://unpkg.com/browse/animated-banner@${packageInfo.version}/dist/animations/${this.data.animation}.js'></script>`
   }
   public copy(text): void {
     copyToClipboard(text)