+++ src/app/gui/show-code-dialog/tabs/cdn-tab.component.ts
@@ -1,6 +1,7 @@
 import {Component, Input} from '@angular/core'
-import {Params} from "../../../model/params";
+import {ComponentParams} from "../../../model/component-params";
 import {copyToClipboard} from "../../../shared/utils";
+import packageInfo  from '../../../../../package.json'
 
 @Component({
   selector: 'app-cdn-tab',
@@ -8,12 +9,12 @@ import {copyToClipboard} from "../../../shared/utils";
 })
 export class CdnTabComponent {
 
-  @Input() data: Params
+  @Input() elementId: string
+  @Input() data: ComponentParams
   public stencilUrl: string = "https://stenciljs.com/docs/javascript"
-  public link1: string = "<script src='https://unpkg.com/browse/animated-banner@0.0.10/dist/index.js'></script>"
-  public link1a: string = "<script src='https://unpkg.com/browse/animated-banner@0.0.10/'></script>"
+  public link1: string = `<script src='https://unpkg.com/browse/animated-banner@${packageInfo.version}/dist/index.js'></script>`
   public get link2(): string {
-    return `<script type='module' src='https://unpkg.com/browse/animated-banner@0.0.10/dist/animations/${this.data.animation}.js'></script>`
+    return `<script type='module' src='https://unpkg.com/browse/animated-banner@${packageInfo.version}/dist/animations/${this.data.animation}.js'></script>`
   }
 
   public copy(text): void {