+++ src/assets/animations/fade-in-out.js
@@ -1,20 +1,16 @@
 /**
- * Animation FadeInOut 0.0.2
+ * Animation FadeInOut 0.1.1
  */
-window.addEventListener('fade-in-out', (event) => {
-  abFadeInOut(event.detail.canvas, event.detail.params);
-  console.log('Animation started: fade-in-out');
-});
-console.log('Animation registered: fade-in-out');
-export function abFadeInOut(canvas, params) {
-  const animator = new ABFadeInOut(canvas, params);
-  animator.animate();
-}
+// TODO: maybe fade in out to another color as an option
 class ABFadeInOut {
-  constructor(canvas, params) {
+  constructor(canvas, text, params) {
+    var _a, _b, _c;
     this.canvasWidth = 0;
     this.canvasHeight = 0;
-    // private loop
+    this.text = 'Hello Banner!';
+    this.textposition = 'bottom';
+    this.color = 'white';
+    this.fontsize = 18;
     this.defaultTextX = 0;
     this.textX = 0;
     this.textY = 0;
@@ -25,15 +21,56 @@ class ABFadeInOut {
     this.canvasWidth = this.canvas.clientWidth;
     this.canvasHeight = this.canvas.clientHeight;
     this.ctx = this.canvas.getContext('2d');
-    this.text = params.text;
-    this.textposition = params.textposition;
-    this.color = params.color;
-    this.fontsize = params.fontsize;
-    // this.loop = params.loop
+    this.text = text;
+    this.textposition = (_a = params.textposition) !== null && _a !== void 0 ? _a : this.textposition;
+    this.color = (_b = params.color) !== null && _b !== void 0 ? _b : this.color;
+    this.fontsize = (_c = params.fontsize) !== null && _c !== void 0 ? _c : this.fontsize;
     this.defaultTextX = this.canvas.width / 2;
     this.textX = this.defaultTextX;
     this.textY = this.getTextY();
   }
+  static registerAnimator() {
+    window.addEventListener('fade-in-out', (event) => {
+      const canvas = event.detail.canvas;
+      const text = event.detail.text;
+      const params = event.detail.params;
+      const animator = new ABFadeInOut(canvas, text, params);
+      animator.animate();
+      console.log('Animation started: fade-in-out');
+    });
+    console.log('Animation registered: fade-in-out');
+  }
+  static registerParamsProvider() {
+    window.addEventListener('get-fade-in-out-params', () => {
+      console.log('Got request for: get-fade-in-out-params');
+      // return animation params when editor requests them
+      // 'selected-animation-params' is constant event name for all animations which editor listens
+      window.dispatchEvent(new CustomEvent('selected-animation-params', {
+        detail: {
+          color: {
+            label: 'Text color',
+            type: 'color',
+            default: '#666'
+          },
+          textposition: {
+            label: 'Text position',
+            type: 'select',
+            default: 'bottom',
+            options: ['top', 'middle', 'bottom']
+          },
+          fontsize: {
+            label: 'Font size',
+            type: 'number',
+            default: 18,
+            min: 1,
+            max: 100
+          },
+        }
+      }));
+    });
+    console.log('ParamsProvider registered: fade-in-out');
+  }
+  //implements AnimatorInterface
   animate() {
     this.setup();
     this.runAnimation();
@@ -80,3 +117,6 @@ class ABFadeInOut {
     }
   }
 }
+ABFadeInOut.registerAnimator();
+ABFadeInOut.registerParamsProvider();
+export {};